+++ lib/openapi_first/app.rb
@@ -5,18 +5,18 @@ require 'rack'
 module OpenapiFirst
   class App
     def initialize(
-      app,
+      parent_app,
       spec,
-      namespace:,
-      allow_unknown_operation: !app.nil?
+      namespace:
     )
       @stack = Rack::Builder.app do
         freeze_app
         use OpenapiFirst::Router,
             spec: spec,
-            allow_unknown_operation: allow_unknown_operation
+            namespace: namespace,
+            parent_app: parent_app
         use OpenapiFirst::RequestValidation
-        run OpenapiFirst::OperationResolver.new(app, namespace: namespace)
+        run OpenapiFirst::OperationResolver.new
       end
     end