+++ lib/openapi_first/request_validation.rb
@@ -3,25 +3,21 @@
 require 'rack'
 require 'json_schemer'
 require 'multi_json'
-require_relative 'query_parameters'
 require_relative 'validation_format'
 
 module OpenapiFirst
-  class RequestValidation
-    def initialize(app, options = {})
+  class RequestValidation # rubocop:disable Metrics/ClassLength
+    def initialize(app, _options = {})
       @app = app
-      @allow_unknown_query_parameters = options.fetch(
-        :allow_unknown_query_parameters, false
-      )
     end
 
     def call(env) # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
       operation = env[OpenapiFirst::OPERATION]
       return @app.call(env) unless operation
 
-      req = Rack::Request.new(env)
       catch(:halt) do
-        validate_query_parameters!(env, operation, req.params)
+        validate_query_parameters!(env, operation, env[PARAMS])
+        req = Rack::Request.new(env)
         content_type = req.content_type
         return @app.call(env) unless operation.request_body
 
@@ -32,6 +28,8 @@ module OpenapiFirst
         @app.call(env)
       end
     end
+
+    private
 
     def halt(response)
       throw :halt, response
@@ -83,10 +81,10 @@ module OpenapiFirst
       ).finish
     end
 
-    def request_body_schema(content_type, endpoint)
-      return unless endpoint
+    def request_body_schema(content_type, operation)
+      return unless operation
 
-      endpoint.request_body.content[content_type]&.fetch('schema')
+      operation.request_body.content[content_type]&.fetch('schema')
     end
 
     def serialize_request_body_errors(validation_errors)
@@ -100,27 +98,25 @@ module OpenapiFirst
     end
 
     def validate_query_parameters!(env, operation, params)
-      json_schema = QueryParameters.new(
-        operation: operation,
-        allow_unknown_parameters: @allow_unknown_query_parameters
-      ).to_json_schema
-
+      json_schema = operation.parameters_json_schema
       return unless json_schema
 
+      params = filtered_params(json_schema, params)
       errors = JSONSchemer.schema(json_schema).validate(params)
       if errors.any?
         halt error_response(400, serialize_query_parameter_errors(errors))
       end
-      env[QUERY_PARAMS] = allowed_params(json_schema, params)
+      env[PARAMS] = params
     end
 
-    def allowed_params(json_schema, params)
+    def filtered_params(json_schema, params)
       json_schema['properties']
-        .keys
-        .each_with_object({}) do |parameter_name, filtered|
+        .each_with_object({}) do |key_value, result|
+          parameter_name, schema = key_value
           next unless params.key?(parameter_name)
 
-          filtered[parameter_name] = params[parameter_name]
+          value = params[parameter_name]
+          result[parameter_name] = parse_parameter(value, schema)
         end
     end
 
@@ -130,6 +126,19 @@ module OpenapiFirst
           source: { parameter: File.basename(error['data_pointer']) }
         }.update(ValidationFormat.error_details(error))
       end
+    end
+
+    def parse_parameter(value, schema)
+      return filtered_params(schema, value) if schema['properties']
+
+      begin
+        return Integer(value, 10) if schema['type'] == 'integer'
+        return Float(value) if schema['type'] == 'number'
+      rescue ArgumentError
+        value
+      end
+
+      value
     end
   end
 end